[Fix] 컬럼수정, 삭제 새로운 컬럼 생성 네트워크 오류 해결
