+++ src/components/common/Column.tsx
@@ -1,13 +1,11 @@
 import { useRef, useCallback, Dispatch, SetStateAction, useState } from "react";
 import SortableCard from "@/components/common/SortableCard";
-import { ColumnData } from "@/types/column";
 import {
   SortableContext,
   verticalListSortingStrategy,
 } from "@dnd-kit/sortable";
 import CreateCard from "../ModalContents/CreateCard.modal";
 import ManageColumnModal from "../ModalContents/ManageColumnModal";
-import useDashboardStates from "@/hooks/useDashboardStates";
 import { DetailContent } from "./ModalPopup";
 import EditCard from "../ModalContents/EditCard.modal";
 import CardCount from "./Chip/CardCount.chip";
@@ -15,12 +13,14 @@ import ColorChip from "./Button/ColorChipSmall";
 
 interface ColumnProps {
   column: ColumnData;
-  onEditCardClick?: (card: Card) => void;
+  setIsEditColumn: Dispatch<SetStateAction<boolean>>;
 }
-const Column: React.FC<ColumnProps> = ({ column, onEditCardClick }) => {
-  const states = useDashboardStates();
+const Column: React.FC<ColumnProps> = ({ column, setIsEditColumn }) => {
   const [isCardEdit, setIsCardEdit] = useState(false);
+  const [currentTitle, setCurrentTitle] = useState(column.title);
+  const [isDeleted, setIsDeleted] = useState(false);
   const { id: columnId, title, cards } = column;
+
   const observer = useRef<IntersectionObserver | null>(null);
   const lastCardRef = useCallback(
     (node: HTMLDivElement | null) => {
@@ -34,53 +34,63 @@ const Column: React.FC<ColumnProps> = ({ column, onEditCardClick }) => {
     [cards.length]
   );
   return (
-    <div className="w-full laptop:w-[354px] tablet:max-w-[584px] bg-gray-100 px-[12px] py-[16px] flex flex-col items-center border-r border-gray-200">
-      <div className="flex w-full items-center justify-between mb-[24px]  h-[22px]">
-        <div className="flex items-center gap-3 justify-between w-fit h-[20px]">
-          <ColorChip color="#5534da" />
-          <p className="text-black-400 text-lg-bold mr-[12px]">{title}</p>
-          <CardCount count={cards.length} />
-        </div>
-        <ManageColumnModal setTitle={states.setTargetColumnTitle} />
-      </div>
+    <>
+      {isDeleted ? null : (
+        <div className="w-full laptop:w-[354px] tablet:max-w-[584px] bg-gray-100 px-[12px] py-[16px] flex flex-col items-center border-r border-gray-200">
+          <div className="flex w-full items-center justify-between mb-[24px]  h-[22px]">
+            <div className="flex items-center gap-3 justify-between w-fit h-[20px]">
+              <ColorChip color="#5534da" />
+              <p className="text-black-400 text-lg-bold mr-[12px]">
+                {currentTitle}
+              </p>
+              <CardCount count={cards.length} />
+            </div>
+            <ManageColumnModal
+              setCurrentTitle={setCurrentTitle}
+              columnId={column.id}
+              title={currentTitle}
+              setIsDeleted={setIsDeleted}
+            />
+          </div>
 
-      <CreateCard columnId={columnId} />
-      <div className="flex w-full flex-col gap-[16px]">
-        <SortableContext
-          items={cards.map((card) => card.cardId)}
-          strategy={verticalListSortingStrategy}
-        >
-          {cards.map((card, index) => (
-            <>
-              <EditCard
-                setIsCardEdit={setIsCardEdit}
-                isCardEdit={isCardEdit}
-                cardId={card.cardId}
-              />
-              <DetailContent
-                columnTitle={states.targetCardColumnTitle}
-                cardId={card.cardId}
-                columnId={card.columnId}
-                cardTitle={card.title}
-                ModalOpenButton={
-                  <SortableCard
-                    key={`card-${card.cardId ?? index}`}
-                    card={card}
-                    columnId={columnId}
-                    index={index}
-                    onClick={() => onEditCardClick?.(card)}
-                    lastCardRef={
-                      index === cards.length - 1 ? lastCardRef : undefined
-                    }
+          <CreateCard columnId={columnId} />
+          <div className="flex w-full flex-col gap-[16px]">
+            <SortableContext
+              items={cards.map((card) => card.cardId)}
+              strategy={verticalListSortingStrategy}
+            >
+              {cards.map((card, index) => (
+                <>
+                  <EditCard
+                    setIsCardEdit={setIsCardEdit}
+                    isCardEdit={isCardEdit}
+                    cardId={card.cardId}
                   />
-                }
-                setIsCardEdit={setIsCardEdit}
-              ></DetailContent>
-            </>
-          ))}
-        </SortableContext>
-      </div>
-    </div>
+                  <DetailContent
+                    columnTitle={column.title}
+                    cardId={card.cardId}
+                    columnId={card.columnId}
+                    cardTitle={card.title}
+                    ModalOpenButton={
+                      <SortableCard
+                        key={`card-${card.cardId ?? index}`}
+                        card={card}
+                        columnId={columnId}
+                        index={index}
+                        lastCardRef={
+                          index === cards.length - 1 ? lastCardRef : undefined
+                        }
+                      />
+                    }
+                    setIsCardEdit={setIsCardEdit}
+                  ></DetailContent>
+                </>
+              ))}
+            </SortableContext>
+          </div>
+        </div>
+      )}
+    </>
   );
 };
