+++ src/components/common/SortableColumn.tsx
@@ -1,22 +1,14 @@
 import { useSortable } from "@dnd-kit/sortable";
 import { CSS } from "@dnd-kit/utilities";
 import Column from "./Column";
-import type { ColumnData } from "@/types/column";
 import { Dispatch, SetStateAction } from "react";
 
 interface Props {
   column: ColumnData;
-  onAddCardClick: (columnId: number) => void;
-  onCardClick: (card: Card) => void;
   setIsEditColumn: Dispatch<SetStateAction<boolean>>;
 }
 
-const SortableColumn = ({
-  column,
-  onAddCardClick,
-  setIsEditColumn,
-  onCardClick: onEditCardClick,
-}: Props) => {
+const SortableColumn = ({ column, setIsEditColumn }: Props) => {
   const { attributes, listeners, setNodeRef, transform, transition } =
     useSortable({
       id: column.id,
@@ -29,7 +21,7 @@ const SortableColumn = ({
 
   return (
     <div ref={setNodeRef} style={style} {...attributes} {...listeners}>
-      <Column column={column} onEditCardClick={onEditCardClick} />
+      <Column column={column} setIsEditColumn={setIsEditColumn} />
     </div>
   );
 };
