+++ src/hooks/useFetchColumns.tsx
@@ -1,15 +1,15 @@
 import { getColumns } from "@/api/column.api";
 import { getCards } from "@/api/card.api";
 import { getDashboardInfo } from "@/api/dashboard";
+import { useStore } from "zustand";
+import useAuthStore from "@/utils/Zustand/zustand";
 
 export function useFetchColumns(setColumns: any, setIsLoading: any) {
+  const authStore = useStore(useAuthStore);
+  const dashboardId = Number(authStore.dashboardId);
   const fetchColumns = async (pageId: number) => {
     try {
       setIsLoading(true);
-
-      const dashboardInfo = await getDashboardInfo(pageId);
-      const dashboardId = dashboardInfo.id;
-
       const columnList = await getColumns(dashboardId);
 
       const fullColumns = await Promise.all(
