--- src/types/column.ts
@@ -1,3 +1,0 @@
-export interface ColumnData extends Column {
-  cards: Card[];
-}
